Big changes to framerate and updates
Framerate and updates are now uncapped.
The amount each entity is updated by is dependant on the amount of time
passed since the last time they were updated.  This is so that all
devices still run the game at the same speed reguardless of the number
of updates / frames rendered.

By using this new updating system, I was able to determine the new way
of rendering the baskground was, in fact, more efficient. As such, it
has been re-added to the game.